feat(cqrs): added mem_store module
