+++ crates/cqrs/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod aggregate;
+pub mod mem_store;
 pub mod store;
